LayoutTests/platform/chromium-win/http/tests/security/xss-DENIED-history-properties-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8080/security/resources/xss-DENIED-history-properties-attacker.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-history-properties.html. Domains, protocols and ports must match.


PASS
